refactor: domain/review -> domain/locationReview 로 패키지 네이밍 변경
