+++ src/main/java/com/workat/api/review/service/ReviewService.java
@@ -17,14 +17,14 @@ import com.workat.api.review.dto.ReviewWithUserDto;
 import com.workat.api.review.dto.request.ReviewRequest;
 import com.workat.common.exception.BadRequestException;
 import com.workat.common.exception.NotFoundException;
+import com.workat.domain.locationReview.entity.BaseReview;
+import com.workat.domain.locationReview.entity.CafeReview;
+import com.workat.domain.locationReview.entity.RestaurantReview;
+import com.workat.domain.locationReview.repository.CafeReviewRepository;
+import com.workat.domain.locationReview.repository.RestaurantReviewRepository;
 import com.workat.domain.map.entity.Location;
 import com.workat.domain.map.entity.LocationCategory;
 import com.workat.domain.map.repository.location.LocationRepository;
-import com.workat.domain.review.entity.BaseReview;
-import com.workat.domain.review.entity.CafeReview;
-import com.workat.domain.review.entity.RestaurantReview;
-import com.workat.domain.review.repository.CafeReviewRepository;
-import com.workat.domain.review.repository.RestaurantReviewRepository;
 import com.workat.domain.tag.BaseTag;
 import com.workat.domain.tag.CafeReviewType;
 import com.workat.domain.tag.FoodReviewType;
