+++ src/main/java/com/workat/domain/locationReview/repository/CafeReviewRepository.java
@@ -1,11 +1,11 @@
-package com.workat.domain.review.repository;
+package com.workat.domain.locationReview.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.workat.domain.review.entity.CafeReview;
+import com.workat.domain.locationReview.entity.CafeReview;
 
 public interface CafeReviewRepository extends JpaRepository<CafeReview, Long> {
 	List<CafeReview> findAllByLocation_Id(long locationId);
