+++ src/main/java/com/workat/domain/locationReview/repository/RestaurantReviewRepository.java
@@ -1,11 +1,11 @@
-package com.workat.domain.review.repository;
+package com.workat.domain.locationReview.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.workat.domain.review.entity.RestaurantReview;
+import com.workat.domain.locationReview.entity.RestaurantReview;
 
 public interface RestaurantReviewRepository extends JpaRepository<RestaurantReview, Long> {
 	List<RestaurantReview> findAllByLocation_Id(long locationId);
