+++ src/test/java/com/workat/api/map/service/LocationServiceTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -25,14 +24,15 @@ import org.springframework.web.client.RestTemplate;
 
 import com.workat.api.map.dto.LocationDetailDto;
 import com.workat.api.map.dto.request.LocationTriggerRequest;
-import com.workat.api.map.dto.response.LocationResponse;
 import com.workat.api.review.service.ReviewService;
 import com.workat.common.exception.BadRequestException;
 import com.workat.common.exception.NotFoundException;
-import com.workat.domain.auth.OauthType;
 import com.workat.domain.area.repository.AreaRepository;
+import com.workat.domain.auth.OauthType;
 import com.workat.domain.config.DataJpaTestConfig;
 import com.workat.domain.config.MysqlContainerBaseTest;
+import com.workat.domain.locationReview.repository.CafeReviewRepository;
+import com.workat.domain.locationReview.repository.RestaurantReviewRepository;
 import com.workat.domain.map.entity.Location;
 import com.workat.domain.map.entity.LocationCategory;
 import com.workat.domain.map.http.LocationHttpReceiver;
@@ -40,8 +40,6 @@ import com.workat.domain.map.http.dto.KakaoLocalDataDto;
 import com.workat.domain.map.http.dto.KakaoLocalMetaDto;
 import com.workat.domain.map.http.dto.KakaoLocalResponse;
 import com.workat.domain.map.repository.location.LocationRepository;
-import com.workat.domain.review.repository.CafeReviewRepository;
-import com.workat.domain.review.repository.RestaurantReviewRepository;
 import com.workat.domain.user.entity.UserProfile;
 import com.workat.domain.user.entity.Users;
 import com.workat.domain.user.job.DepartmentType;
