+++ src/test/java/com/workat/api/review/service/ReviewServiceTest.java
@@ -26,14 +26,14 @@ import com.workat.common.exception.BadRequestException;
 import com.workat.domain.auth.OauthType;
 import com.workat.domain.config.DataJpaTestConfig;
 import com.workat.domain.config.MysqlContainerBaseTest;
+import com.workat.domain.locationReview.entity.CafeReview;
+import com.workat.domain.locationReview.entity.RestaurantReview;
+import com.workat.domain.locationReview.repository.CafeReviewRepository;
+import com.workat.domain.locationReview.repository.RestaurantReviewRepository;
 import com.workat.domain.map.entity.Location;
 import com.workat.domain.map.entity.LocationCategory;
 import com.workat.domain.map.http.dto.KakaoLocalDataDto;
 import com.workat.domain.map.repository.location.LocationRepository;
-import com.workat.domain.review.entity.CafeReview;
-import com.workat.domain.review.entity.RestaurantReview;
-import com.workat.domain.review.repository.CafeReviewRepository;
-import com.workat.domain.review.repository.RestaurantReviewRepository;
 import com.workat.domain.tag.CafeReviewType;
 import com.workat.domain.user.entity.UserProfile;
 import com.workat.domain.user.entity.Users;
